refactor: Infrastructure層のディレクトリ構成を変更.
