+++ internal/app/api/infrastructure/db/policy_test.go
@@ -5,7 +5,7 @@ import (
 	"database/sql"
 	"fmt"
 	"holos-auth-api/internal/app/api/domain/entity"
-	"holos-auth-api/internal/app/api/infrastructure"
+	infrastructure "holos-auth-api/internal/app/api/infrastructure/db"
 	"holos-auth-api/internal/app/api/pkg/apierr"
 	"holos-auth-api/test"
 	"net/http"
@@ -17,7 +17,7 @@ import (
 	"github.com/google/uuid"
 )
 
-func TestAgent_Create(t *testing.T) {
+func TestPolicy_Create(t *testing.T) {
 	tests := []struct {
 		name          string
 		isTransaction bool
@@ -33,7 +33,7 @@ func TestAgent_Create(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			agent, err := entity.NewAgent(uuid.New(), "name")
+			policy, err := entity.NewPolicy(uuid.New(), "name", "STORAGE", "/", []string{"GET"})
 			if err != nil {
 				t.Error(err.Error())
 			}
@@ -46,23 +46,23 @@ func TestAgent_Create(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectExec(regexp.QuoteMeta("INSERT INTO agents (id, user_id, name, created_at, updated_at) VALUES (?, ?, ?, ?, ?);")).
-				WithArgs(agent.ID, agent.UserID, agent.Name, agent.CreatedAt, agent.UpdatedAt).
+			mock.ExpectExec(regexp.QuoteMeta("INSERT INTO policies (id, user_id, name, service, path, allowed_methods, created_at, updated_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?);")).
+				WithArgs(policy.ID, policy.UserID, policy.Name, policy.Service, policy.Path, `["GET"]`, policy.CreatedAt, policy.UpdatedAt).
 				WillReturnResult(sqlmock.NewResult(1, 1))
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			ai := infrastructure.NewAgentInfrastructure(db)
+			pi := infrastructure.NewPolicyInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					return ai.Create(ctx, agent)
+					return pi.Create(ctx, policy)
 				}); err != nil {
 					t.Error(err.Error())
 				}
 			} else {
-				if err := ai.Create(ctx, agent); err != nil {
+				if err := pi.Create(ctx, policy); err != nil {
 					t.Error(err.Error())
 				}
 			}
@@ -70,7 +70,7 @@ func TestAgent_Create(t *testing.T) {
 	}
 }
 
-func TestAgent_Update(t *testing.T) {
+func TestPolicy_Update(t *testing.T) {
 	tests := []struct {
 		name          string
 		isTransaction bool
@@ -86,7 +86,7 @@ func TestAgent_Update(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			agent, err := entity.NewAgent(uuid.New(), "name")
+			policy, err := entity.NewPolicy(uuid.New(), "name", "STORAGE", "/", []string{"GET"})
 			if err != nil {
 				t.Error(err.Error())
 			}
@@ -99,23 +99,23 @@ func TestAgent_Update(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectExec(regexp.QuoteMeta("UPDATE agents SET user_id = ?, name = ?, updated_at = ? WHERE id = ? AND deleted_at IS NULL LIMIT 1;")).
-				WithArgs(agent.UserID, agent.Name, agent.UpdatedAt, agent.ID).
+			mock.ExpectExec(regexp.QuoteMeta("UPDATE policies SET user_id = ?, name = ?, service = ?, path = ?, allowed_methods = ?, updated_at = ? WHERE id = ? AND deleted_at IS NULL LIMIT 1;")).
+				WithArgs(policy.UserID, policy.Name, policy.Service, policy.Path, `["GET"]`, policy.UpdatedAt, policy.ID).
 				WillReturnResult(sqlmock.NewResult(1, 1))
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			ai := infrastructure.NewAgentInfrastructure(db)
+			pi := infrastructure.NewPolicyInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					return ai.Update(ctx, agent)
+					return pi.Update(ctx, policy)
 				}); err != nil {
 					t.Error(err.Error())
 				}
 			} else {
-				if err := ai.Update(ctx, agent); err != nil {
+				if err := pi.Update(ctx, policy); err != nil {
 					t.Error(err.Error())
 				}
 			}
@@ -123,7 +123,7 @@ func TestAgent_Update(t *testing.T) {
 	}
 }
 
-func TestAgent_Delete(t *testing.T) {
+func TestPolicy_Delete(t *testing.T) {
 	tests := []struct {
 		name          string
 		isTransaction bool
@@ -139,7 +139,7 @@ func TestAgent_Delete(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			agent, err := entity.NewAgent(uuid.New(), "name")
+			policy, err := entity.NewPolicy(uuid.New(), "name", "STORAGE", "/", []string{"GET"})
 			if err != nil {
 				t.Error(err.Error())
 			}
@@ -152,23 +152,23 @@ func TestAgent_Delete(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectExec(regexp.QuoteMeta("UPDATE agents SET updated_at = updated_at, deleted_at = NOW(6) WHERE id = ? AND deleted_at IS NULL LIMIT 1;")).
-				WithArgs(agent.ID).
+			mock.ExpectExec(regexp.QuoteMeta("UPDATE policies SET updated_at = updated_at, deleted_at = NOW(6) WHERE id = ? AND deleted_at IS NULL LIMIT 1;")).
+				WithArgs(policy.ID).
 				WillReturnResult(sqlmock.NewResult(1, 1))
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			ai := infrastructure.NewAgentInfrastructure(db)
+			pi := infrastructure.NewPolicyInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					return ai.Delete(ctx, agent)
+					return pi.Delete(ctx, policy)
 				}); err != nil {
 					t.Error(err.Error())
 				}
 			} else {
-				if err := ai.Delete(ctx, agent); err != nil {
+				if err := pi.Delete(ctx, policy); err != nil {
 					t.Error(err.Error())
 				}
 			}
@@ -176,7 +176,7 @@ func TestAgent_Delete(t *testing.T) {
 	}
 }
 
-func TestAgent_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
+func TestPolicy_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 	tests := []struct {
 		id            uuid.UUID
 		userID        uuid.UUID
@@ -220,22 +220,22 @@ func TestAgent_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectQuery(regexp.QuoteMeta("SELECT id, user_id, name, created_at, updated_at FROM agents WHERE id = ? AND user_id = ? AND deleted_at IS NULL LIMIT 1;")).
+			mock.ExpectQuery(regexp.QuoteMeta("SELECT id, user_id, name, service, path, allowed_methods, created_at, updated_at FROM policies WHERE id = ? AND user_id = ? AND deleted_at IS NULL LIMIT 1;")).
 				WithArgs(tt.id, tt.userID).
 				WillReturnRows(
-					sqlmock.NewRows([]string{"id", "user_id", "name", "created_at", "updated_at"}).
-						AddRow(tt.id, tt.userID, tt.name, time.Now(), time.Now()),
+					sqlmock.NewRows([]string{"id", "user_id", "name", "service", "path", "allowed_methods", "created_at", "updated_at"}).
+						AddRow(tt.id, tt.userID, tt.name, "STORAGE", "/", `["GET"]`, time.Now(), time.Now()),
 				).
 				WillReturnError(tt.resultError)
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			ai := infrastructure.NewAgentInfrastructure(db)
+			pi := infrastructure.NewPolicyInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					result, err := ai.FindOneByIDAndUserIDAndNotDeleted(ctx, tt.id, tt.userID)
+					result, err := pi.FindOneByIDAndUserIDAndNotDeleted(ctx, tt.id, tt.userID)
 					if err != nil {
 						return err
 					}
@@ -247,7 +247,7 @@ func TestAgent_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 					t.Error(err.Error())
 				}
 			} else {
-				result, err := ai.FindOneByIDAndUserIDAndNotDeleted(ctx, tt.id, tt.userID)
+				result, err := pi.FindOneByIDAndUserIDAndNotDeleted(ctx, tt.id, tt.userID)
 				if err != nil {
 					t.Error(err.Error())
 				}
@@ -259,7 +259,7 @@ func TestAgent_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 	}
 }
 
-func TestAgent_FindByUserIDAndNotDeleted(t *testing.T) {
+func TestPolicy_FindByUserIDAndNotDeleted(t *testing.T) {
 	tests := []struct {
 		id            uuid.UUID
 		userID        uuid.UUID
@@ -295,22 +295,22 @@ func TestAgent_FindByUserIDAndNotDeleted(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectQuery(regexp.QuoteMeta("SELECT id, user_id, name, created_at, updated_at FROM agents WHERE user_id = ? AND deleted_at IS NULL;")).
+			mock.ExpectQuery(regexp.QuoteMeta("SELECT id, user_id, name, service, path, allowed_methods, created_at, updated_at FROM policies WHERE user_id = ? AND deleted_at IS NULL;")).
 				WithArgs(tt.userID).
 				WillReturnRows(
-					sqlmock.NewRows([]string{"id", "user_id", "name", "created_at", "updated_at"}).
-						AddRow(tt.id, tt.userID, tt.name, time.Now(), time.Now()),
+					sqlmock.NewRows([]string{"id", "user_id", "name", "service", "path", "allowed_methods", "created_at", "updated_at"}).
+						AddRow(tt.id, tt.userID, tt.name, "STORAGE", "/", `["GET"]`, time.Now(), time.Now()),
 				).
 				WillReturnError(tt.resultError)
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			ai := infrastructure.NewAgentInfrastructure(db)
+			pi := infrastructure.NewPolicyInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					result, err := ai.FindByUserIDAndNotDeleted(ctx, tt.userID)
+					result, err := pi.FindByUserIDAndNotDeleted(ctx, tt.userID)
 					if err != nil {
 						return err
 					}
@@ -322,7 +322,7 @@ func TestAgent_FindByUserIDAndNotDeleted(t *testing.T) {
 					t.Error(err.Error())
 				}
 			} else {
-				result, err := ai.FindByUserIDAndNotDeleted(ctx, tt.userID)
+				result, err := pi.FindByUserIDAndNotDeleted(ctx, tt.userID)
 				if err != nil {
 					t.Error(err.Error())
 				}
