+++ internal/app/api/infrastructure/db/user_test.go
@@ -5,7 +5,7 @@ import (
 	"database/sql"
 	"fmt"
 	"holos-auth-api/internal/app/api/domain/entity"
-	"holos-auth-api/internal/app/api/infrastructure"
+	infrastructure "holos-auth-api/internal/app/api/infrastructure/db"
 	"holos-auth-api/internal/app/api/pkg/apierr"
 	"holos-auth-api/test"
 	"net/http"
@@ -17,7 +17,7 @@ import (
 	"github.com/google/uuid"
 )
 
-func TestPolicy_Create(t *testing.T) {
+func TestUser_Create(t *testing.T) {
 	tests := []struct {
 		name          string
 		isTransaction bool
@@ -33,7 +33,7 @@ func TestPolicy_Create(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			policy, err := entity.NewPolicy(uuid.New(), "name", "STORAGE", "/", []string{"GET"})
+			user, err := entity.NewUser(tt.name, "password", "password")
 			if err != nil {
 				t.Error(err.Error())
 			}
@@ -46,23 +46,23 @@ func TestPolicy_Create(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectExec(regexp.QuoteMeta("INSERT INTO policies (id, user_id, name, service, path, allowed_methods, created_at, updated_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?);")).
-				WithArgs(policy.ID, policy.UserID, policy.Name, policy.Service, policy.Path, `["GET"]`, policy.CreatedAt, policy.UpdatedAt).
+			mock.ExpectExec(regexp.QuoteMeta("INSERT INTO users (id, name, password, created_at, updated_at) VALUES (?, ?, ?, ?, ?);")).
+				WithArgs(user.ID, user.Name, user.Password, user.CreatedAt, user.UpdatedAt).
 				WillReturnResult(sqlmock.NewResult(1, 1))
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			pi := infrastructure.NewPolicyInfrastructure(db)
+			ui := infrastructure.NewUserInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					return pi.Create(ctx, policy)
+					return ui.Create(ctx, user)
 				}); err != nil {
 					t.Error(err.Error())
 				}
 			} else {
-				if err := pi.Create(ctx, policy); err != nil {
+				if err := ui.Create(ctx, user); err != nil {
 					t.Error(err.Error())
 				}
 			}
@@ -70,7 +70,7 @@ func TestPolicy_Create(t *testing.T) {
 	}
 }
 
-func TestPolicy_Update(t *testing.T) {
+func TestUser_Update(t *testing.T) {
 	tests := []struct {
 		name          string
 		isTransaction bool
@@ -86,7 +86,7 @@ func TestPolicy_Update(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			policy, err := entity.NewPolicy(uuid.New(), "name", "STORAGE", "/", []string{"GET"})
+			user, err := entity.NewUser(tt.name, "password", "password")
 			if err != nil {
 				t.Error(err.Error())
 			}
@@ -99,23 +99,23 @@ func TestPolicy_Update(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectExec(regexp.QuoteMeta("UPDATE policies SET user_id = ?, name = ?, service = ?, path = ?, allowed_methods = ?, updated_at = ? WHERE id = ? AND deleted_at IS NULL LIMIT 1;")).
-				WithArgs(policy.UserID, policy.Name, policy.Service, policy.Path, `["GET"]`, policy.UpdatedAt, policy.ID).
+			mock.ExpectExec(regexp.QuoteMeta("UPDATE users SET name = ?, password = ?, updated_at = ? WHERE id = ? AND deleted_at IS NULL LIMIT 1;")).
+				WithArgs(user.Name, user.Password, user.UpdatedAt, user.ID).
 				WillReturnResult(sqlmock.NewResult(1, 1))
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			pi := infrastructure.NewPolicyInfrastructure(db)
+			ui := infrastructure.NewUserInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					return pi.Update(ctx, policy)
+					return ui.Update(ctx, user)
 				}); err != nil {
 					t.Error(err.Error())
 				}
 			} else {
-				if err := pi.Update(ctx, policy); err != nil {
+				if err := ui.Update(ctx, user); err != nil {
 					t.Error(err.Error())
 				}
 			}
@@ -123,7 +123,7 @@ func TestPolicy_Update(t *testing.T) {
 	}
 }
 
-func TestPolicy_Delete(t *testing.T) {
+func TestUser_Delete(t *testing.T) {
 	tests := []struct {
 		name          string
 		isTransaction bool
@@ -139,7 +139,7 @@ func TestPolicy_Delete(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			policy, err := entity.NewPolicy(uuid.New(), "name", "STORAGE", "/", []string{"GET"})
+			user, err := entity.NewUser(tt.name, "password", "password")
 			if err != nil {
 				t.Error(err.Error())
 			}
@@ -152,23 +152,35 @@ func TestPolicy_Delete(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectExec(regexp.QuoteMeta("UPDATE policies SET updated_at = updated_at, deleted_at = NOW(6) WHERE id = ? AND deleted_at IS NULL LIMIT 1;")).
-				WithArgs(policy.ID).
+			mock.ExpectExec(regexp.QuoteMeta(
+				`UPDATE users
+				LEFT JOIN agents ON users.id = agents.user_id
+				SET
+					users.updated_at = users.updated_at,
+					users.deleted_at = NOW(6),
+					agents.updated_at = agents.updated_at,
+					agents.deleted_at = NOW(6)
+				WHERE
+					users.id = ?
+					AND users.deleted_at IS NULL
+					AND agents.deleted_at IS NULL;`,
+			)).
+				WithArgs(user.ID).
 				WillReturnResult(sqlmock.NewResult(1, 1))
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			pi := infrastructure.NewPolicyInfrastructure(db)
+			ui := infrastructure.NewUserInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					return pi.Delete(ctx, policy)
+					return ui.Delete(ctx, user)
 				}); err != nil {
 					t.Error(err.Error())
 				}
 			} else {
-				if err := pi.Delete(ctx, policy); err != nil {
+				if err := ui.Delete(ctx, user); err != nil {
 					t.Error(err.Error())
 				}
 			}
@@ -176,10 +188,9 @@ func TestPolicy_Delete(t *testing.T) {
 	}
 }
 
-func TestPolicy_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
+func TestUser_FindOneByIDAndNotDeleted(t *testing.T) {
 	tests := []struct {
 		id            uuid.UUID
-		userID        uuid.UUID
 		name          string
 		isTransaction bool
 		resultIsNil   bool
@@ -187,7 +198,6 @@ func TestPolicy_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 	}{
 		{
 			id:            uuid.New(),
-			userID:        uuid.New(),
 			name:          "without_transaction",
 			isTransaction: false,
 			resultIsNil:   false,
@@ -195,7 +205,6 @@ func TestPolicy_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 		},
 		{
 			id:            uuid.New(),
-			userID:        uuid.New(),
 			name:          "with_transaction",
 			isTransaction: true,
 			resultIsNil:   false,
@@ -203,8 +212,7 @@ func TestPolicy_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 		},
 		{
 			id:            uuid.New(),
-			userID:        uuid.New(),
-			name:          "agent_not_found",
+			name:          "user_not_found",
 			isTransaction: false,
 			resultIsNil:   true,
 			resultError:   sql.ErrNoRows,
@@ -220,22 +228,22 @@ func TestPolicy_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectQuery(regexp.QuoteMeta("SELECT id, user_id, name, service, path, allowed_methods, created_at, updated_at FROM policies WHERE id = ? AND user_id = ? AND deleted_at IS NULL LIMIT 1;")).
-				WithArgs(tt.id, tt.userID).
+			mock.ExpectQuery(regexp.QuoteMeta("SELECT id, name, password, created_at, updated_at FROM users WHERE id = ? AND deleted_at IS NULL LIMIT 1;")).
+				WithArgs(tt.id).
 				WillReturnRows(
-					sqlmock.NewRows([]string{"id", "user_id", "name", "service", "path", "allowed_methods", "created_at", "updated_at"}).
-						AddRow(tt.id, tt.userID, tt.name, "STORAGE", "/", `["GET"]`, time.Now(), time.Now()),
+					sqlmock.NewRows([]string{"id", "name", "password", "created_at", "updated_at"}).
+						AddRow(tt.id, tt.name, "password", time.Now(), time.Now()),
 				).
 				WillReturnError(tt.resultError)
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			pi := infrastructure.NewPolicyInfrastructure(db)
+			ui := infrastructure.NewUserInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					result, err := pi.FindOneByIDAndUserIDAndNotDeleted(ctx, tt.id, tt.userID)
+					result, err := ui.FindOneByIDAndNotDeleted(ctx, tt.id)
 					if err != nil {
 						return err
 					}
@@ -247,7 +255,7 @@ func TestPolicy_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 					t.Error(err.Error())
 				}
 			} else {
-				result, err := pi.FindOneByIDAndUserIDAndNotDeleted(ctx, tt.id, tt.userID)
+				result, err := ui.FindOneByIDAndNotDeleted(ctx, tt.id)
 				if err != nil {
 					t.Error(err.Error())
 				}
@@ -259,30 +267,30 @@ func TestPolicy_FindOneByIDAndUserIDAndNotDeleted(t *testing.T) {
 	}
 }
 
-func TestPolicy_FindByUserIDAndNotDeleted(t *testing.T) {
+func TestUser_FindOneByName(t *testing.T) {
 	tests := []struct {
-		id            uuid.UUID
-		userID        uuid.UUID
 		name          string
 		isTransaction bool
 		resultIsNil   bool
 		resultError   error
 	}{
 		{
-			id:            uuid.New(),
-			userID:        uuid.New(),
 			name:          "without_transaction",
 			isTransaction: false,
 			resultIsNil:   false,
 			resultError:   nil,
 		},
 		{
-			id:            uuid.New(),
-			userID:        uuid.New(),
 			name:          "with_transaction",
 			isTransaction: true,
 			resultIsNil:   false,
 			resultError:   nil,
+		},
+		{
+			name:          "user_not_found",
+			isTransaction: false,
+			resultIsNil:   true,
+			resultError:   sql.ErrNoRows,
 		},
 	}
 	for _, tt := range tests {
@@ -295,22 +303,22 @@ func TestPolicy_FindByUserIDAndNotDeleted(t *testing.T) {
 			if tt.isTransaction {
 				mock.ExpectBegin()
 			}
-			mock.ExpectQuery(regexp.QuoteMeta("SELECT id, user_id, name, service, path, allowed_methods, created_at, updated_at FROM policies WHERE user_id = ? AND deleted_at IS NULL;")).
-				WithArgs(tt.userID).
+			mock.ExpectQuery(regexp.QuoteMeta("SELECT id, name, password, created_at, updated_at FROM users WHERE name = ? LIMIT 1;")).
+				WithArgs(tt.name).
 				WillReturnRows(
-					sqlmock.NewRows([]string{"id", "user_id", "name", "service", "path", "allowed_methods", "created_at", "updated_at"}).
-						AddRow(tt.id, tt.userID, tt.name, "STORAGE", "/", `["GET"]`, time.Now(), time.Now()),
+					sqlmock.NewRows([]string{"id", "name", "password", "created_at", "updated_at"}).
+						AddRow(uuid.New(), tt.name, "password", time.Now(), time.Now()),
 				).
 				WillReturnError(tt.resultError)
 			if tt.isTransaction {
 				mock.ExpectCommit()
 			}
 
-			pi := infrastructure.NewPolicyInfrastructure(db)
+			ui := infrastructure.NewUserInfrastructure(db)
 			if tt.isTransaction {
 				to := infrastructure.NewSqlxTransactionObject(db)
 				if err := to.Transaction(ctx, func(ctx context.Context) apierr.ApiError {
-					result, err := pi.FindByUserIDAndNotDeleted(ctx, tt.userID)
+					result, err := ui.FindOneByName(ctx, tt.name)
 					if err != nil {
 						return err
 					}
@@ -322,7 +330,7 @@ func TestPolicy_FindByUserIDAndNotDeleted(t *testing.T) {
 					t.Error(err.Error())
 				}
 			} else {
-				result, err := pi.FindByUserIDAndNotDeleted(ctx, tt.userID)
+				result, err := ui.FindOneByName(ctx, tt.name)
 				if err != nil {
 					t.Error(err.Error())
 				}
